# docs/swagger.yaml
definitions:
...
  github_com_PritOriginal_problem-map-server_internal_models.AdminBoundary:
    properties:
      admin_level:
        type: integer
      geom:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.MultiPolygonJSON'
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_PritOriginal_problem-map-server_internal_models.AdminBoundaryMarksCount:
    properties:
      closed_count:
        type: integer
      confirmed_count:
        type: integer
      id:
        type: integer
      name:
        type: string
      total_count:
        type: integer
      unconfirmed_count:
        type: integer
      under_review_count:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.City:
    properties:
      city_id:
        type: integer
      geom:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.PolygonJSON'
      name:
        type: string
      region_id:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.District:
    properties:
      city_id:
        type: integer
      district_id:
        type: integer
      geom:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.PolygonJSON'
      name:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.MultiPolygonJSON:
    properties:
      coordinates:
        items:
          items:
            items:
              items:
                type: number
              type: array
            type: array
          type: array
        type: array
      type:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.PolygonJSON:
    properties:
      coordinates:
        items:
          items:
            type: number
          type: array
        type: array
      type:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.Region:
    properties:
      geom:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.PolygonJSON'
      name:
        type: string
      region_id:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo:
    properties:
      message:
        type: string
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-any:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload: {}
      success:
        type: boolean
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetAdminBoundariesMarksCountResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.GetAdminBoundariesMarksCountResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetAdminBoundariesResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.GetAdminBoundariesResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetCitiesResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.GetCitiesResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetDistrictsResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.GetDistrictsResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetRegionsResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.GetRegionsResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_HeatmapResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_map.HeatmapResponse'
      success:
        type: boolean
    type: object
...
  internal_handler_map.AdminBoundaryFeature:
    properties:
      geometry:
        type: object
      id:
        type: integer
      properties:
        $ref: '#/definitions/internal_handler_map.AdminBoundaryProperties'
      type:
        example: Feature
        type: string
    type: object
  internal_handler_map.AdminBoundaryProperties:
    properties:
      admin_level:
        type: integer
      name:
        type: string
    type: object
  internal_handler_map.GetAdminBoundariesMarksCountResponse:
    properties:
      admin_boundaries:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.AdminBoundaryMarksCount'
        type: array
    type: object
  internal_handler_map.GetAdminBoundariesResponse:
    properties:
      admin_boundaries:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.AdminBoundary'
        type: array
    type: object
  internal_handler_map.GetCitiesResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.City'
        type: array
    type: object
  internal_handler_map.GetDistrictsResponse:
    properties:
      districts:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.District'
        type: array
    type: object
  internal_handler_map.GetRegionsResponse:
    properties:
      regions:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.Region'
        type: array
    type: object
  internal_handler_map.HeatmapFeature:
    properties:
      geometry:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.PolygonJSON'
      properties:
        $ref: '#/definitions/internal_handler_map.HeatmapProperties'
      type:
        enum:
        - Feature
        type: string
    type: object
  internal_handler_map.HeatmapProperties:
    properties:
      count:
        type: integer
    type: object
  internal_handler_map.HeatmapResponse:
    properties:
      features:
        items:
          $ref: '#/definitions/internal_handler_map.HeatmapFeature'
        type: array
      type:
        enum:
        - FeatureCollection
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the "Problem Map" project.
  title: Problem Map API
  version: "1.0"
paths:
...
  /map/admin-boundaries:
    get:
      consumes:
      - application/json
      description: admin boundaries
      parameters:
      - collectionFormat: csv
        description: filter by admin level
        in: query
        items:
          type: number
        name: admin_levels
        type: array
      - description: ETag of a previous response; 304 when the data did not change
        in: header
        name: If-None-Match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: public, max-age=60
              type: string
            ETag:
              description: validator for If-None-Match
              type: string
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetAdminBoundariesResponse'
        "304":
          description: not modified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List administrative boundaries
      tags:
      - map
  /map/admin-boundaries/{id}.geojson:
    get:
      description: 'one boundary with its MultiPolygon geometry as a GeoJSON Feature
        (`application/geo+json`): `properties` carry `name` and `admin_level`'
      parameters:
      - description: boundary id (the path is /map/admin-boundaries/{id}.geojson)
        in: path
        name: id
        required: true
        type: integer
      - description: ETag of a previous response; 304 when the boundary did not change
        in: header
        name: If-None-Match
        type: string
      produces:
      - application/geo+json
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: public, max-age=86400
              type: string
            ETag:
              description: validator for If-None-Match
              type: string
          schema:
            $ref: '#/definitions/internal_handler_map.AdminBoundaryFeature'
        "304":
          description: not modified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Administrative boundary as GeoJSON
      tags:
      - map
  /map/admin-boundaries/marks/count:
    get:
      consumes:
      - application/json
      description: the count of markers of all administrative boundaries
      parameters:
      - collectionFormat: csv
        description: filter by admin level
        in: query
        items:
          type: number
        name: admin_levels
        type: array
      - collectionFormat: csv
        description: filter by mark type
        in: query
        items:
          type: number
        name: mark_type_ids
        type: array
      - collectionFormat: csv
        description: filter by mark status
        in: query
        items:
          type: number
        name: mark_status_ids
        type: array
      - description: marks created at or after (RFC3339 or YYYY-MM-DD, taken as 00:00:00
          UTC)
        in: query
        name: from
        type: string
      - description: marks created at or before (RFC3339 or YYYY-MM-DD, taken as the
          end of that day in UTC)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetAdminBoundariesMarksCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: The count of markers of all administrative boundaries
      tags:
      - map
  /map/cities:
    get:
      consumes:
      - application/json
      description: get cities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetCitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List cities
      tags:
      - map
  /map/districts:
    get:
      consumes:
      - application/json
      description: get districts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetDistrictsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List districts
      tags:
      - map
  /map/heatmap:
    get:
      description: 'GeoJSON FeatureCollection of hexagons (EPSG:3857 grid, returned
        in WGS84) with the number of marks in each; empty cells are omitted. At most
        5000 cells: a finer grid is rejected with 400, increase cell_m. Cached for
        60 seconds per query.'
      parameters:
      - description: minLon,minLat,maxLon,maxLat
        in: query
        name: bbox
        required: true
        type: string
      - default: 250
        description: hexagon size (center-to-vertex) in ground meters (10..100000)
        in: query
        name: cell_m
        type: number
      - collectionFormat: csv
        description: filter by mark type
        in: query
        items:
          type: number
        name: mark_type_ids
        type: array
      - collectionFormat: csv
        description: filter by mark status
        in: query
        items:
          type: number
        name: mark_status_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_HeatmapResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Marks heatmap
      tags:
      - map
  /map/regions:
    get:
      consumes:
      - application/json
      description: get regions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_map_GetRegionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List regions
      tags:
      - map
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Open-data API key "pm_live_…" (POST /api-keys); also accepted as
      "Authorization: ApiKey {key}". Optional on public GET routes: with a key the
      per-key quota applies and X-RateLimit-* headers are returned; ignored when a
      valid Bearer JWT is sent as well (the JWT identity wins)'
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT access token: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...